pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "sunflare"
dynamic = ["version"]
description = "Redsun plugin development toolkit"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Jacopo Abramo", email = "jacopo.abramo@gmail.com"}
]
maintainers = [
  {name = "Jacopo Abramo", email = "jacopo.abramo@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
    "License :: OSI Approved :: Apache Software License",
]
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml",
    "types-PyYAML",
    "bluesky",
    "attrs",
    "psygnal",
    "numpy>=1.24, <1.27",
    "typing_extensions>=4.10.0",
    "zmq",
    "msgspec",
]

[project.optional-dependencies]
pyqt = ["qtpy", "pyqt6"]
pyside = ["qtpy", "pyside6"]
docs = [
    "sphinx",
    "pydata-sphinx-theme",
    "sphinxcontrib-mermaid",
    "sphinx-design",
    "myst-parser"
]
dev = [
  "ruff",
  "pytest",
  "pytest-cov",
  "pre-commit",
  "ophyd",
  "mypy"
]

[project.urls]
bugs = "https://github.com/redsun-acquisition/sunflare/issues"
changelog = "https://redsun-acquisition.github.io/sunflare/changelog/"
homepage = "https://github.com/redsun-acquisition/sunflare"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/sunflare"]
exclude = [
    "src/sunflare/docs",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/sunflare/*.py",
  "src/sunflare/docs/**",
  "tests/**",
]

[tool.mypy]
files = "src/**/*.py"
exclude = ["docs", "tests", "build"]
disable_error_code = ["import-untyped", "no-untyped-call"]
pretty = true

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
disallow_untyped_defs = true

[tool.coverage.run]
source = ["sunflare"]
omit = [
    "tests/conftest.py",
    "src/sunflare/view/**",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]

# https://coverage.readthedocs.io/en/latest/config.html
[tool.coverage.report]
exclude_lines = [
    "@abstractmethod",
    "if TYPE_CHECKING:",
    "@overload",
    "pragma: no cover",
    "except ImportError",
    "except PackageNotFoundError",
    "raise NotImplementedError()",
]

[tool.ruff]
target-version = "py39"
extend-exclude = ["docs"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
extend-select = [
    "D", # enable docstring checks
    "I", # enable imports checks
    "TC", # enable type check imports
    "FA102", # force __future__ imports
] 
ignore = ["D100", "D104"] # https://docs.astral.sh/ruff/rules/undocumented-public-package/
exclude = ["tests/**"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
